+++ specs/resources/stock_items.spec.ts
@@ -328,4 +328,30 @@ describe('StockItems resource', () => {
 	/* relationship.versions stop */
 	
   
+
+	/* trigger._validate start */
+	it(resourceType + '._validate', async () => {
+	
+		let triggerAttr = '_validate'
+		if (!triggerAttr.startsWith('_')) triggerAttr = `_${triggerAttr}`
+	
+		const triggerValue = true
+		const attributes = { [triggerAttr]: triggerValue }
+	    const id = TestData.id
+	
+		const intId = cl.addRequestInterceptor((request) => {
+			const data = JSON.parse(String(request.options.body))
+			expect(request.options.method).toBe('PATCH')
+			checkCommon(request, resourceType, id, currentAccessToken)
+			checkCommonData(data, resourceType, attributes, id)
+			return interceptRequest()
+		})
+	
+		await cl[resourceType]._validate(id, {}, CommonData.options)
+			.catch(handleError)
+			.finally(() => cl.removeInterceptor('request'))
+	
+	})
+	/* trigger._validate stop */
+	
 })