+++ src/resources/paypal_payments.ts
@@ -65,21 +65,6 @@ interface PaypalPayment extends Resource {
 	 */
 	mismatched_amounts?: boolean | null
 	/** 
-	 * The amount of the associated payment intent, in cents..
-	 * @example ```"1000"```
-	 */
-	intent_amount_cents: number
-	/** 
-	 * The amount of the associated payment intent, float..
-	 * @example ```"10"```
-	 */
-	intent_amount_float?: number | null
-	/** 
-	 * The amount of the associated payment intent, formatted..
-	 * @example ```"€10,00"```
-	 */
-	formatted_intent_amount?: string | null
-	/** 
 	 * Information about the payment instrument used in the transaction.
 	 * @example ```"[object Object]"```
 	 */