+++ src/resources/satispay_payments.ts
@@ -46,21 +46,6 @@ interface SatispayPayment extends Resource {
 	 */
 	payment_url?: string | null
 	/** 
-	 * The amount of the associated payment intent, in cents..
-	 * @example ```"1000"```
-	 */
-	intent_amount_cents: number
-	/** 
-	 * The amount of the associated payment intent, float..
-	 * @example ```"10"```
-	 */
-	intent_amount_float?: number | null
-	/** 
-	 * The amount of the associated payment intent, formatted..
-	 * @example ```"€10,00"```
-	 */
-	formatted_intent_amount?: string | null
-	/** 
 	 * The Satispay payment response, used to fetch internal data..
 	 * @example ```"[object Object]"```
 	 */